Databricks/fine-tuning/requirements-databricks.txt
# Databricks BERT Fine-tuning Requirements
# Install these packages in your Databricks cluster or notebook

# Core ML frameworks
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0
datasets>=2.0.0

# MLflow (usually pre-installed in Databricks)
mlflow>=2.5.0

# Databricks specific
databricks-sdk>=0.8.0

# AWS integration
boto3>=1.26.0
s3fs>=2023.1.0

# Utilities
pyyaml>=6.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: Additional optimization libraries
# apex  # For advanced mixed precision training (requires manual installation)
# deepspeed>=0.9.0  # For distributed training optimization

# Development and debugging
ipykernel>=6.0.0
jupyter>=1.0.0

# Note: The following are typically pre-installed in Databricks Runtime ML:
# - PySpark
# - Koalas/pandas API on Spark  
# - Hyperopt
# - MLflow
# - Delta Lake
# - Unity Catalog integration
